Displaying password after generating. Fixing modes.
Presenting the standard output from the `pass generate` command
instead of loading the new password by default.

Adding a 5 second timeout and error handling in case the
`pass generate` command hangs.

Fixing a bug where "j" and "k" keys are handled as vim-like
key-bindings on on the screen for generating a new password.
Using new modes: `generate` and `generate_results`.

Fixing a typo on the generate password screen.

Adding myself to the AUTHORS file.